feat: 添加 Apple Silicon MPS 支持
- 添加 MPS 设备检测和适配
- 修复 float32 兼容性问题
- 添加 run_mps_inference.py 推理脚本
- 修复 scipy 导入问题

+++ m2a/espnet2/layers/global_mvn.py
@@ -53,8 +53,9 @@ class GlobalMVN(AbsNormalize, InversibleInterface):
             var = sum_square_v / count - mean * mean
         std = np.sqrt(np.maximum(var, eps))
 
-        self.register_buffer("mean", torch.from_numpy(mean))
-        self.register_buffer("std", torch.from_numpy(std))
+        # Use float32 for MPS compatibility (MPS doesn't support float64)
+        self.register_buffer("mean", torch.from_numpy(mean.astype(np.float32)))
+        self.register_buffer("std", torch.from_numpy(std.astype(np.float32)))
 
     def extra_repr(self):
         return (
